feat(schedule-meeting): implement AvailableSchedule component for time slot selection
