+++ src/components/huber-registration/Step2.tsx
@@ -1,8 +1,6 @@
 'use client';
 
 import { zodResolver } from '@hookform/resolvers/zod';
-import { Box } from '@mui/material';
-import clsx from 'clsx';
 import { useTranslations } from 'next-intl';
 import { useState } from 'react';
 import { Controller, useForm } from 'react-hook-form';
@@ -12,73 +10,8 @@ import Button from '@/components/button/Button';
 import HeadSlots from '@/components/common/HeadSlots';
 import { useCreateTimeSlotsMutation } from '@/libs/services/modules/time-slots';
 
+import AvailableSchedule from '../common/AvailableSchedule';
 import { pushError, pushSuccess } from '../CustomToastifyContainer';
-
-const timeSlots = {
-  morning: [
-    { time: '06:00', day: 0 },
-    { time: '06:30', day: 0 },
-    { time: '07:00', day: 0 },
-    { time: '07:30', day: 0 },
-    { time: '08:00', day: 0 },
-    { time: '08:30', day: 0 },
-    { time: '09:00', day: 0 },
-    { time: '09:30', day: 0 },
-    { time: '10:00', day: 0 },
-    { time: '10:30', day: 0 },
-    { time: '11:00', day: 0 },
-    { time: '11:30', day: 0 },
-  ],
-  afternoon: [
-    { time: '12:00', day: 0 },
-    { time: '12:30', day: 0 },
-    { time: '13:00', day: 0 },
-    { time: '13:30', day: 0 },
-    { time: '14:00', day: 0 },
-    { time: '14:30', day: 0 },
-    { time: '15:00', day: 0 },
-    { time: '15:30', day: 0 },
-    { time: '16:00', day: 0 },
-    { time: '16:30', day: 0 },
-    { time: '17:00', day: 0 },
-    { time: '17:30', day: 0 },
-  ],
-  evening: [
-    { time: '18:00', day: 0 },
-    { time: '18:30', day: 0 },
-    { time: '19:00', day: 0 },
-    { time: '19:30', day: 0 },
-    { time: '20:00', day: 0 },
-    { time: '20:30', day: 0 },
-    { time: '21:00', day: 0 },
-    { time: '21:30', day: 0 },
-    { time: '22:00', day: 0 },
-    { time: '22:30', day: 0 },
-    { time: '23:00', day: 0 },
-    { time: '23:30', day: 0 },
-  ],
-};
-
-// Helper function to generate time slots for all 7 days
-const generateTimeSlotsForWeek = () => {
-  const daysOfWeek = [0, 1, 2, 3, 4, 5, 6]; // 0 = Sunday, 1 = Monday, ..., 6 = Saturday
-  const result: typeof timeSlots = { morning: [], afternoon: [], evening: [] };
-
-  daysOfWeek.forEach((day) => {
-    Object.entries(timeSlots).forEach(([period, slots]) => {
-      slots.forEach((slot) => {
-        result[period as keyof typeof timeSlots].push({
-          ...slot,
-          day,
-        });
-      });
-    });
-  });
-
-  return result;
-};
-
-const timeSlotsForWeek = generateTimeSlotsForWeek();
 
 const Step2 = ({ next }: { next: () => void }) => {
   const t = useTranslations('HumanBookRegister.Step2');
@@ -171,42 +104,14 @@ const Step2 = ({ next }: { next: () => void }) => {
                   <HeadSlots
                     dayOfWeek={currentDay}
                     onChangeDayOfWeek={setCurrentDay}
+                    selectedTimes={field.value}
                   />
 
-                  {Object.entries(timeSlotsForWeek).map(([period, slots]) => {
-                    const filteredSlots = slots.filter(
-                      (slot) => slot.day === currentDay,
-                    );
-                    return (
-                      <div
-                        key={period}
-                        className="flex flex-col gap-2 rounded-lg bg-neutral-98 p-2"
-                      >
-                        <div className="flex flex-wrap justify-between gap-1">
-                          {filteredSlots.map((slot) => (
-                            <Box
-                              key={`${slot.day}-${slot.time}`}
-                              onClick={() =>
-                                handleTimeSelect(slot.time, slot.day)
-                              }
-                              className={clsx(
-                                'w-[70px] cursor-pointer rounded-full py-1 text-center text-sm',
-                                field.value.some(
-                                  (s) =>
-                                    s.startTime === slot.time &&
-                                    s.dayOfWeek === slot.day,
-                                )
-                                  ? 'bg-green-90 text-green-30'
-                                  : 'bg-gray-200 text-gray-700 hover:bg-gray-300 ',
-                              )}
-                            >
-                              {slot.time}
-                            </Box>
-                          ))}
-                        </div>
-                      </div>
-                    );
-                  })}
+                  <AvailableSchedule
+                    onSelectTime={handleTimeSelect}
+                    selectedTimes={field.value}
+                    currentDay={currentDay}
+                  />
                 </>
               );
             }}
